🚚 Rename: 직관적인 네이밍으로 변경
