+++ src/pages/profile/Profile.tsx
@@ -1,4 +1,4 @@
-import { ActivityLog } from '@widgets/profile/components'
+import { ActivityLogSection } from '@widgets/profile/components'
 import { ProfileProps } from './Profile.type'
 import {
   MyProfileCard,
@@ -12,7 +12,7 @@ export default function Profile({ userId }: ProfileProps) {
     <main className="flex justify-center gap-[60px] w-full mt-[160px] tablet:flex-col tablet:px-28 mobile:mt-[130px] mobile:px-4">
       {isMyProfile ? <MyProfileCard /> : <UserProfileCard userId={userId} />}
       <div className="flex flex-col gap-[60px]">
-        <ActivityLog userId={userId} />
+        <ActivityLogSection userId={userId} />
         <section className="col-start-2 tablet:col-start-1">카드자리</section>
       </div>
     </main>
