+++ src/widgets/profile/components/activity-log-section/index.ts
@@ -1,3 +1,3 @@
 export { default as LogList } from './list/LogList'
 export { default as LogListItem } from './item/LogListItem'
-export { default as ActivityLog } from './ActivityLog'
+export { default as ActivityLog } from './ActivityLogSection'
