+++ src/widgets/profile/components/activity-log-section/item/LogListItem.type.ts
@@ -1,5 +1,3 @@
-import { FavoriteCategory } from '@shared/types'
-
 export interface LogListItemProps {
   title: string
   image?: string
