--- src/widgets/profile/components/image-section/ProfileImageSection.type.ts
@@ -1,7 +1,0 @@
-import type { UserItem } from '@shared/types'
-
-export interface ProfileImageSectionProps {
-  image: string | null
-  description: string
-  nickname: string
-}
