+++ src/widgets/profile/components/profile-header/ProfileHeader.type.ts
@@ -1,0 +1,5 @@
+export interface ProfileHeaderProps {
+  image: string | null
+  description: string
+  nickname: string
+}
